+++ src/ts/components/button.ts
@@ -27,16 +27,29 @@ export class Button<Config extends ButtonConfig> extends Component<Config> {
 
     this.config = this.mergeConfig(config, {
       cssClass: 'ui-button',
+      role: 'button',
+      tabIndex: 0,
     } as Config, this.config);
   }
 
   protected toDomElement(): DOM {
-    // Create the button element with the text label
-    let buttonElement = new DOM('button', {
-      'type': 'button',
+    const buttonElementAttributes: { [name: string]: string } = {
       'id': this.config.id,
+      'aria-label': i18n.performLocalization(this.config.ariaLabel || this.config.text),
       'class': this.getCssClasses(),
-    }).append(new DOM('span', {
+      /**
+      * WCAG20 standard to display if a button is pressed or not
+      */
+      'aria-pressed': 'false',
+      'tabindex': this.config.tabIndex.toString(),
+    };
+
+    if (this.config.role != null) {
+      buttonElementAttributes['role'] = this.config.role;
+    }
+
+    // Create the button element with the text label
+    let buttonElement = new DOM('button', buttonElementAttributes).append(new DOM('span', {
       'class': this.prefixCss('label'),
     }).html(i18n.performLocalization(this.config.text)));