+++ src/ts/components/seekbar.ts
@@ -9,6 +9,9 @@ import TimeShiftAvailabilityChangedArgs = PlayerUtils.TimeShiftAvailabilityChang
 import LiveStreamDetectorEventArgs = PlayerUtils.LiveStreamDetectorEventArgs;
 import { TimelineMarker } from '../uiconfig';
 import { PlayerAPI, PlayerEventBase } from 'bitmovin-player';
+import { StringUtils } from '../stringutils';
+import { SeekBarType, SeekBarController } from './seekbarcontroller';
+import { i18n } from '../localization/i18n';
 
 /**
  * Configuration interface for the {@link SeekBar} component.
@@ -30,6 +33,11 @@ export interface SeekBarConfig extends ComponentConfig {
    * Default: 50 (50ms = 20fps).
    */
   smoothPlaybackPositionUpdateIntervalMs?: number;
+
+  /**
+   * Used for seekBar control increments and decrements
+   */
+  keyStepIncrements?: { leftRight: number, upDown: number };
 }
 
 /**
@@ -80,6 +88,10 @@ export class SeekBar extends Component<SeekBarConfig> {
 
   private player: PlayerAPI;
 
+  protected seekBarType: SeekBarType;
+
+  protected isUiShown: boolean;
+
   /**
    * Buffer of the the current playback position. The position must be buffered in case the element
    * needs to be refreshed with {@link #refreshPlaybackPosition}.
@@ -111,10 +123,18 @@ export class SeekBar extends Component<SeekBarConfig> {
   constructor(config: SeekBarConfig = {}) {
     super(config);
 
+    const keyStepIncrements = this.config.keyStepIncrements || {
+      leftRight: 1,
+      upDown: 5,
+    };
+
     this.config = this.mergeConfig(config, {
       cssClass: 'ui-seekbar',
       vertical: false,
       smoothPlaybackPositionUpdateIntervalMs: 50,
+      keyStepIncrements,
+      ariaLabel: i18n.getLocalizer('seekBar'),
+      tabIndex: 0,
     }, this.config);
 
     this.label = this.config.label;
@@ -129,6 +149,23 @@ export class SeekBar extends Component<SeekBarConfig> {
     }
   }
 
+  protected setAriaSliderMinMax(min: string, max: string) {
+    this.getDomElement().attr('aria-valuemin', min);
+    this.getDomElement().attr('aria-valuemax', max);
+  }
+
+  private setAriaSliderValues() {
+    if (this.seekBarType === SeekBarType.Live) {
+      const timeshiftValue = Math.ceil(this.player.getTimeShift()).toString();
+      this.getDomElement().attr('aria-valuenow', timeshiftValue);
+      this.getDomElement().attr('aria-valuetext', `${i18n.performLocalization(i18n.getLocalizer('seekBar.timeshift'))} ${i18n.performLocalization(i18n.getLocalizer('seekBar.value'))}: ${timeshiftValue}`);
+    } else if (this.seekBarType === SeekBarType.Vod) {
+      const ariaValueText = `${StringUtils.secondsToText(this.player.getCurrentTime())} ${i18n.performLocalization(i18n.getLocalizer('seekBar.durationText'))} ${StringUtils.secondsToText(this.player.getDuration())}`;
+      this.getDomElement().attr('aria-valuenow', Math.floor(this.player.getCurrentTime()).toString());
+      this.getDomElement().attr('aria-valuetext', ariaValueText);
+    }
+  }
+
   configure(player: PlayerAPI, uimanager: UIInstanceManager, configureSeek: boolean = true): void {
     super.configure(player, uimanager);
 
@@ -138,13 +175,28 @@ export class SeekBar extends Component<SeekBarConfig> {
     // (the call must be up here to be executed for the volume slider as well)
     this.setPosition(this.seekBarBackdrop, 100);
 
+    // Add seekbar controls to the seekbar
+    const seekBarController = new SeekBarController(this.config.keyStepIncrements, player, uimanager.getConfig().volumeController);
+
+    seekBarController.setSeekBarControls(this.getDomElement(), () => this.seekBarType);
+
+    // The configureSeek flag can be used by subclasses to disable configuration as seek bar. E.g. the volume
+    // slider is reusing this component but adds its own functionality, and does not need the seek functionality.
+    // This is actually a hack, the proper solution would be for both seek bar and volume sliders to extend
+    // a common base slider component and implement their functionality there.
     if (!configureSeek) {
-      // The configureSeek flag can be used by subclasses to disable configuration as seek bar. E.g. the volume
-      // slider is reusing this component but adds its own functionality, and does not need the seek functionality.
-      // This is actually a hack, the proper solution would be for both seek bar and volume sliders to extend
-      // a common base slider component and implement their functionality there.
+      this.seekBarType = SeekBarType.Volume;
+
       return;
     }
+
+    uimanager.onControlsShow.subscribe(() => {
+      this.isUiShown = true;
+    });
+
+    uimanager.onControlsHide.subscribe(() => {
+      this.isUiShown = false;
+    });
 
     let isPlaying = false;
     let isUserSeeking = false;
@@ -164,12 +216,14 @@ export class SeekBar extends Component<SeekBarConfig> {
         } else {
           let playbackPositionPercentage = 100 - (100 / player.getMaxTimeShift() * player.getTimeShift());
           this.setPlaybackPosition(playbackPositionPercentage);
+          this.setAriaSliderMinMax(player.getMaxTimeShift().toString(), '0');
         }
 
         // Always show full buffer for live streams
         this.setBufferPosition(100);
       } else {
-        let playbackPositionPercentage = 100 / player.getDuration() * this.getRelativeCurrentTime();
+        const playerDuration = player.getDuration();
+        let playbackPositionPercentage = 100 / playerDuration * this.getRelativeCurrentTime();
 
         let videoBufferLength = player.getVideoBufferLength();
         let audioBufferLength = player.getAudioBufferLength();
@@ -184,7 +238,7 @@ export class SeekBar extends Component<SeekBarConfig> {
           bufferLength = 0;
         }
 
-        let bufferPercentage = 100 / player.getDuration() * bufferLength;
+        let bufferPercentage = 100 / playerDuration * bufferLength;
 
         // Update playback position only in paused state or in the initial startup state where player is neither
         // paused nor playing. Playback updates are handled in the Timeout below.
@@ -194,6 +248,12 @@ export class SeekBar extends Component<SeekBarConfig> {
         }
 
         this.setBufferPosition(playbackPositionPercentage + bufferPercentage);
+
+        this.setAriaSliderMinMax('0', playerDuration.toString());
+      }
+
+      if (this.isUiShown) {
+        this.setAriaSliderValues();
       }
     };
 
@@ -297,6 +357,9 @@ export class SeekBar extends Component<SeekBarConfig> {
       isLive = args.live;
       if (isLive && this.smoothPlaybackPositionUpdater != null) {
         this.smoothPlaybackPositionUpdater.clear();
+        this.seekBarType = SeekBarType.Live;
+      } else {
+        this.seekBarType = SeekBarType.Vod;
       }
       switchVisibility(isLive, hasTimeShift);
     });
@@ -520,6 +583,9 @@ export class SeekBar extends Component<SeekBarConfig> {
     let seekBarContainer = new DOM('div', {
       'id': this.config.id,
       'class': this.getCssClasses(),
+      'role': 'slider',
+      'aria-label': i18n.performLocalization(this.config.ariaLabel),
+      'tabindex': this.config.tabIndex.toString(),
     });
 
     let seekBar = new DOM('div', {
@@ -578,6 +644,7 @@ export class SeekBar extends Component<SeekBarConfig> {
       this.setPlaybackPosition(targetPercentage);
       this.onSeekPreviewEvent(targetPercentage, true);
     };
+
     let mouseTouchUpHandler = (e: MouseEvent | TouchEvent) => {
       e.preventDefault();