+++ src/ts/components/subtitlesettings/subtitlesettingslabel.ts
@@ -14,11 +14,14 @@ export class SubtitleSettingsLabel extends Container<ContainerConfig> {
 
   private text: LocalizableText;
 
+  private for: string;
+
   constructor(config: SubtitleSettingsLabelConfig) {
     super(config);
 
     this.opener = config.opener;
     this.text = config.text;
+    this.for = config.for;
 
     this.config = this.mergeConfig(<ContainerConfig>config, {
       cssClass: 'ui-label',
@@ -29,9 +32,10 @@ export class SubtitleSettingsLabel extends Container<ContainerConfig> {
   }
 
   protected toDomElement(): DOM {
-    let labelElement = new DOM('span', {
+    let labelElement = new DOM('label', {
       'id': this.config.id,
       'class': this.getCssClasses(),
+      'for': this.for,
     }).append(
       new DOM('span', {}).html(i18n.performLocalization(this.text)),
       this.opener.getDomElement(),