+++ src/ts/dom.ts
@@ -79,7 +79,9 @@ export class DOM {
 
       for (let attributeName in attributes) {
         let attributeValue = attributes[attributeName];
-        element.setAttribute(attributeName, attributeValue);
+        if (attributeValue != null) {
+          element.setAttribute(attributeName, attributeValue);
+        }
       }
 
       this.elements = [element];
@@ -168,6 +170,17 @@ export class DOM {
   }
 
   /**
+   * Focuses to the first input element
+   */
+  focusToFirstInput() {
+    const inputElements = this.findChildElements('button, [href], input, select, textarea, [tabindex]:not([tabindex="-1"])');
+
+    if (inputElements.length > 0) {
+      inputElements[0].focus();
+    }
+  }
+
+  /**
    * Returns a string of the inner HTML content of the first element.
    */
   html(): string;
@@ -248,6 +261,16 @@ export class DOM {
     else {
       return this.getAttr(attribute);
     }
+  }
+
+  /**
+   * Removes the attribute of the element.
+   * @param attribute
+   */
+  removeAttr(attribute: string) {
+    this.forEach((element) => {
+      element.removeAttribute(attribute);
+    });
   }
 
   private getAttr(attribute: string): string | null {