+++ src/ts/uifactory.ts
@@ -84,16 +84,25 @@ export namespace UIFactory {
       overlay: subtitleOverlay,
     });
 
+    const subtitleSelectBox = new SubtitleSelectBox();
+
     let subtitleSettingsOpenButton = new SettingsPanelPageOpenButton({
       targetPage: subtitleSettingsPanelPage,
       container: settingsPanel,
+      ariaLabel: i18n.getLocalizer('settings.subtitles'),
       text: i18n.getLocalizer('open'),
     });
 
     mainSettingsPanelPage.addComponent(
       new SettingsPanelItem(
-        new SubtitleSettingsLabel({text: i18n.getLocalizer('settings.subtitles'), opener: subtitleSettingsOpenButton}),
-        new SubtitleSelectBox(),
+        new SubtitleSettingsLabel({
+          text: i18n.getLocalizer('settings.subtitles'),
+          opener: subtitleSettingsOpenButton,
+        }),
+        subtitleSelectBox,
+        {
+          role: 'menubar',
+        },
       ));
 
     settingsPanel.addComponent(subtitleSettingsPanelPage);
@@ -215,13 +224,22 @@ export namespace UIFactory {
     let subtitleSettingsOpenButton = new SettingsPanelPageOpenButton({
       targetPage: subtitleSettingsPanelPage,
       container: settingsPanel,
+      ariaLabel: i18n.getLocalizer('settings.subtitles'),
       text: i18n.getLocalizer('open'),
     });
 
+    const subtitleSelectBox = new SubtitleSelectBox();
+
     mainSettingsPanelPage.addComponent(
       new SettingsPanelItem(
-        new SubtitleSettingsLabel({text: i18n.getLocalizer('settings.subtitles'), opener: subtitleSettingsOpenButton}),
-        new SubtitleSelectBox(),
+        new SubtitleSettingsLabel({
+          text: i18n.getLocalizer('settings.subtitles'),
+          opener: subtitleSettingsOpenButton,
+        }),
+        subtitleSelectBox,
+        {
+          role: 'menubar',
+        },
       ));
 
     settingsPanel.addComponent(subtitleSettingsPanelPage);